requirements.txt
asgiref==3.3.1
astroid==2.4.2
autopep8==1.5.4
colorama==0.4.4
coverage==5.3
Django==3.1.3
django-debug-toolbar==3.1.1
gunicorn==20.0.4
isort==5.6.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
Pillow==8.0.1
psycopg2==2.8.6
pycodestyle==2.6.0
pylint==2.6.0
python-decouple==3.3
pytz==2020.4
selenium==3.141.0
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.25.11
wrapt==1.12.1
